# apps/web/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AGI Workforce API",
    "version": "1.0.0",
    "description": "API for managing AI agents, subscriptions, and device authorization in AGI Workforce",
    "contact": {
      "name": "AGI Workforce Support",
      "email": "support@agiworkforce.com",
      "url": "https://agiworkforce.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://agiworkforce.com/api",
      "description": "Production API server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key authentication using Bearer token"
      }
    },
    "schemas": {
      "CheckoutRequest": {
        "type": "object",
        "required": ["plan", "billingInterval"],
        "properties": {
          "plan": {
            "type": "string",
            "enum": ["basic", "pro", "max", "max_15x"],
            "description": "The subscription plan to checkout"
          },
          "billingInterval": {
            "type": "string",
            "enum": ["monthly", "yearly"],
            "description": "Billing cycle interval"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/checkout": {
      "post": {
        "summary": "Create a checkout session",
        "description": "Initiate a Stripe checkout session for subscription",
        "tags": ["Billing"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout session created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "format": "uri",
                      "description": "Stripe checkout URL"
                    },
                    "sessionId": {
                      "type": "string",
                      "description": "Stripe session ID"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "401": {
            "description": "Not authenticated"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
  }
}